fourth-year/Software Engineering/Practicals/kasula-backend-develop/app/utils/token.py
import jwt
from datetime import datetime, timedelta
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jwt.exceptions import InvalidTokenError
from typing import Dict, Optional
import os

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token", auto_error=False)

SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 20160  # 14 days

def create_access_token(data: dict, user_id: str, username: str, expires_delta: timedelta = None):
    to_encode = data.copy()
    to_encode["user_id"] = user_id
    to_encode["username"] = username
    to_encode["iat"] = datetime.utcnow()
    
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

def get_current_user(token: str = Depends(oauth2_scheme)) -> Optional[Dict[str, str]]:
    if token is None:
        # No token, user is not authenticated
        return None

    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )

    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        user_id: str = payload.get("user_id")
        username: str = payload.get("username")
        if user_id is None:
            raise credentials_exception
        return {"user_id": user_id, "username": username}
    except (InvalidTokenError, jwt.ExpiredSignatureError):
        raise credentials_exception
